+++ fragnet-search/src/main/java/org/squonk/fragnet/search/model/v2/NeighbourhoodGraph.java
@@ -201,6 +201,7 @@ public class NeighbourhoodGraph extends FragmentGraph {
                     LOG.info("Refmol/MCS Atoms: " + refMolAtoms + "/" + mcsAtoms + " Took: " + (t1 - t0) +
                             "ns Smarts: " + smarts);
                     group.setRefmolAtomsMissing(refMolAtoms - mcsAtoms);
+                    group.setScaffold(smarts);
                 }
             }
         }
@@ -285,7 +286,7 @@ public class NeighbourhoodGraph extends FragmentGraph {
     }
 
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    @JsonPropertyOrder({"key", "classification", "prototype", "refmolAtomsMissing", "members"})
+    @JsonPropertyOrder({"key", "classification", "prototype", "scaffold", "refmolAtomsMissing", "members"})
     public class Group {
 
         private final String key;
@@ -293,6 +294,7 @@ public class NeighbourhoodGraph extends FragmentGraph {
         private String prototype;
         private Integer refmolAtomsMissing;
         private final List<GroupMember> members = new ArrayList<>();
+        private String scaffold;
 
         protected Group(String key) {
             this.key = key;
@@ -317,6 +319,14 @@ public class NeighbourhoodGraph extends FragmentGraph {
 
         protected void setPrototype(String prototype) {
             this.prototype = prototype;
+        }
+
+        public String getScaffold() {
+            return scaffold;
+        }
+
+        public void setScaffold(String scaffold) {
+            this.scaffold = scaffold;
         }
 
         public Integer getRefmolAtomsMissing() {
@@ -503,7 +513,6 @@ public class NeighbourhoodGraph extends FragmentGraph {
 
         protected String generateGroupingKey() {
 
-
             String result = null;
             if (edges.size() == 1) {
                 // only a single route between the query and the member
@@ -534,16 +543,22 @@ public class NeighbourhoodGraph extends FragmentGraph {
                     String s = splitLabel(l[0])[4];
                     parts04.add(s);
                 }
-                // check if all the parts are the same
-                Set<String> set = new HashSet<>(parts04);
-                if (set.size() == 1) {
-                    // this happens when there are multiple edges between the same nodes
-                    // we only want one of them
-                    result = parts04.get(0);
-                } else {
-                    Collections.sort(parts04);
-                    result = parts04.stream().collect(Collectors.joining("$$"));
-                }
+//
+                // remove duplicate paths that come multiple edges between the same nodes due to symmetry
+                // and ensure the paths are sorted
+                Set<String> unique = new TreeSet<>(parts04);
+                result = unique.stream().collect(Collectors.joining("$$"));
+
+//                // check if all the parts are the same
+//                Set<String> set = new HashSet<>(parts04);
+//                if (set.size() == 1) {
+//                    // this happens when there are multiple edges between the same nodes
+//                    // we only want one of them
+//                    result = parts04.get(0);
+//                } else {
+//                    Collections.sort(parts04);
+//                    result = parts04.stream().collect(Collectors.joining("$$"));
+//                }
 
             }
             if (result == null) {