+++ fragnet-search/src/test/groovy/org/squonk/fragnet/search/model/v2/GroupMemberSpec.groovy
@@ -4,6 +4,11 @@ import spock.lang.Specification
 
 class GroupMemberSpec extends Specification {
 
+    static NeighbourhoodGraph.GroupMember createMember(long id, String smiles) {
+        MoleculeNode molNode = new MoleculeNode(id, smiles, MoleculeNode.MoleculeType.NET_MOL, [], [:])
+        new NeighbourhoodGraph.GroupMember(molNode)
+    }
+
     void "one edge between nodes one hop"() {
 
         // This tests the case where there is a single one hop path between a pair on nodes
@@ -89,16 +94,17 @@ class GroupMemberSpec extends Specification {
                         "FG|CC[Xe]|CC[100Xe]|RING|Cc1cc(-c2cccc([Xe])c2)ccc1O|CC1CC(C2CCCC([100Xe])C2)CCC1O")
         ] as MoleculeEdge[])
         def key = member.generateGroupingKey()
-        println key
+        //println key
 
         then:
         key == 'Oc1ccc(-c2cccc([Xe])c2)cc1$$Oc1ccc(-c2ccccc2)cc1[Xe]'
     }
 
+    // this is biphenol substituted at the 3 position
     void "three routes between nodes two hops 35995193"() {
 
         // MATCH p=(m:F2)-[:FRAG*1..2]-(e:F2)
-        // WHERE m.smiles='Oc1ccc(-c2ccccc2)cc1' AND e.smiles='C=CCc1cccc(-c2ccccc2)c1' RETURN p LIMIT 1000
+        // WHERE m.smiles='Oc1ccc(-c2ccccc2)cc1' AND e.smiles='C=CCc1cccc(-c2ccccc2)c1' RETURN p
 
         when:
         NeighbourhoodGraph.GroupMember member = createMember(35995193, "C=CCc1cccc(-c2ccccc2)c1")
@@ -122,15 +128,112 @@ class GroupMemberSpec extends Specification {
                         "FG|O[Xe]|O[102Xe]|RING|C=CCc1cc(-c2ccccc2)ccc1[Xe]|CCCC1CC(C2CCCCC2)CCC1[102Xe]")
         ] as MoleculeEdge[])
         def key = member.generateGroupingKey()
-        println key
+        //println key
 
         then:
         key == 'Oc1ccc(-c2cccc([Xe])c2)cc1$$Oc1ccc(-c2ccccc2)cc1[Xe]$$[Xe]c1ccc(-c2ccccc2)cc1'
     }
 
-    static NeighbourhoodGraph.GroupMember createMember(long id, String smiles) {
-        MoleculeNode molNode = new MoleculeNode(id, smiles, MoleculeNode.MoleculeType.NET_MOL, [], [:])
-        new NeighbourhoodGraph.GroupMember(molNode)
+    // this is biphenol substituted at the 3 position
+    void "two routes between nodes two hops 65424227"() {
+
+        // MATCH p=(m:F2)-[:FRAG*1..2]-(e:F2)
+        // WHERE m.smiles='Oc1ccc(-c2ccccc2)cc1' AND e.smiles='CCc1cccc(-c2ccccc2)c1' RETURN p
+
+        when:
+        NeighbourhoodGraph.GroupMember member = createMember(65424227, "CCc1cccc(-c2ccccc2)c1")
+
+        member.addEdges([
+                new MoleculeEdge(636675733, 123525973, 123867746,
+                        "FG|O[Xe]|O[100Xe]|RING|[Xe]c1ccc(-c2ccccc2)cc1|[100Xe]C1CCC(C2CCCCC2)CC1"),
+                new MoleculeEdge(303615864, 65424227, 123867746,
+                        "FG|CC[Xe]|CC[100Xe]|RING|[Xe]c1cccc(-c2ccccc2)c1|[100Xe]C1CCCC(C2CCCCC2)C1")
+        ] as MoleculeEdge[])
+        member.addEdges([
+                new MoleculeEdge(303613227, 65424086, 123525973,
+                        "FG|CC[Xe]|CC[100Xe]|RING|Oc1ccc(-c2cccc([Xe])c2)cc1|OC1CCC(C2CCCC([100Xe])C2)CC1"),
+                new MoleculeEdge(303613224, 65424086, 65424227,
+                        "FG|O[Xe]|O[102Xe]|RING|CCc1cccc(-c2ccc([Xe])cc2)c1|CCC1CCCC(C2CCC([102Xe])CC2)C1")
+        ] as MoleculeEdge[])
+        def key = member.generateGroupingKey()
+        //println key
+
+        then:
+        key == 'Oc1ccc(-c2cccc([Xe])c2)cc1$$[Xe]c1ccc(-c2ccccc2)cc1'
+    }
+
+    // this is biphenol substituted at the 3 position
+    void "two routes between nodes two hops 111475017"() {
+
+        // MATCH p=(m:F2)-[:FRAG*1..2]-(e:F2)
+        // WHERE m.smiles='Oc1ccc(-c2ccccc2)cc1' AND e.smiles='Ic1cccc(-c2ccccc2)c1'
+
+        when:
+        NeighbourhoodGraph.GroupMember member = createMember(111475017, "Ic1cccc(-c2ccccc2)c1")
+
+        member.addEdges([
+                new MoleculeEdge(636675733, 123525973, 123867746,
+                        "FG|O[Xe]|O[100Xe]|RING|[Xe]c1ccc(-c2ccccc2)cc1|[100Xe]C1CCC(C2CCCCC2)CC1"),
+                new MoleculeEdge(575952673, 111475017, 123867746,
+                        "FG|I[Xe]|I[100Xe]|RING|[Xe]c1cccc(-c2ccccc2)c1|[100Xe]C1CCCC(C2CCCCC2)C1")
+        ] as MoleculeEdge[])
+        member.addEdges([
+                new MoleculeEdge(636680811, 123526164, 123525973,
+                        "FG|I[Xe]|I[102Xe]|RING|Oc1ccc(-c2ccccc2)cc1[Xe]|OC1CCC(C2CCCCC2)CC1[102Xe]"),
+                new MoleculeEdge(636680810, 123526164, 111475017,
+                        "FG|O[Xe]|O[100Xe]|RING|Ic1cc(-c2ccccc2)ccc1[Xe]|IC1CC(C2CCCCC2)CCC1[100Xe]")
+        ] as MoleculeEdge[])
+        def key = member.generateGroupingKey()
+        //println key
+
+        then:
+        key == 'Oc1ccc(-c2ccccc2)cc1[Xe]$$[Xe]c1ccc(-c2ccccc2)cc1'
+    }
+
+    // this is biphenol substituted at the 3 position
+    void "four routes between nodes two hops 114309881"() {
+
+        // case where there are 4 routes, with two of them being duplicate because of symmetry
+        // In this case the second and third sets of edges share the same first edge.
+        // The results should be that the duplicate path is not included so there will
+        // only be 3 parts to the grouping key
+
+        // MATCH p=(m:F2)-[:FRAG*1..2]-(e:F2)
+        // WHERE m.smiles='Oc1ccc(-c2ccccc2)cc1' AND e.smiles='Nc1cccc(-c2ccccc2)c1' RETURN p
+
+        when:
+        NeighbourhoodGraph.GroupMember member = createMember(114309881, "Nc1cccc(-c2ccccc2)c1")
+
+        member.addEdges([
+                new MoleculeEdge(589205488, 114309515, 123525973,
+                        "FG|N[Xe]|N[100Xe]|RING|Oc1ccc(-c2cccc([Xe])c2)cc1|OC1CCC(C2CCCC([100Xe])C2)CC1"),
+                new MoleculeEdge(589205485, 114309515, 114309881,
+                        "FG|O[Xe]|O[102Xe]|RING|Nc1cccc(-c2ccc([Xe])cc2)c1|NC1CCCC(C2CCC([102Xe])CC2)C1")
+        ] as MoleculeEdge[])
+        member.addEdges([
+                new MoleculeEdge(636675733, 123525973, 123867746,
+                        "FG|O[Xe]|O[100Xe]|RING|[Xe]c1ccc(-c2ccccc2)cc1|[100Xe]C1CCC(C2CCCCC2)CC1"),
+                new MoleculeEdge(589221302, 114309881, 123867746,
+                        "FG|N[Xe]|N[101Xe]|RING|[Xe]c1cccc(-c2ccccc2)c1|[101Xe]C1CCCC(C2CCCCC2)C")
+        ] as MoleculeEdge[])
+        member.addEdges([
+                new MoleculeEdge(636675733, 123525973, 123867746,
+                        "FG|O[Xe]|O[100Xe]|RING|[Xe]c1ccc(-c2ccccc2)cc1|[100Xe]C1CCC(C2CCCCC2)CC1"),
+                new MoleculeEdge(589221301, 114309881, 123867746,
+                        "FG|N[Xe]|N[100Xe]|RING|[Xe]c1cccc(-c2ccccc2)c1|[100Xe]C1CCCC(C2CCCCC2)C1")
+        ] as MoleculeEdge[])
+        member.addEdges([
+                new MoleculeEdge(588432788, 114062345, 123525973,
+                        "FG|N[Xe]|N[100Xe]|RING|Oc1ccc(-c2ccccc2)cc1[Xe]|OC1CCC(C2CCCCC2)CC1[100Xe]"),
+                new MoleculeEdge(588432785, 114062345, 114309881,
+                        "FG|O[Xe]|O[102Xe]|RING|Nc1cc(-c2ccccc2)ccc1[Xe]|NC1CC(C2CCCCC2)CCC1[102Xe]")
+        ] as MoleculeEdge[])
+
+        def key = member.generateGroupingKey()
+        //println key
+
+        then:
+        key == 'Oc1ccc(-c2cccc([Xe])c2)cc1$$Oc1ccc(-c2ccccc2)cc1[Xe]$$[Xe]c1ccc(-c2ccccc2)cc1'
     }
 
 }